session: drop and merge cockpitwsinstancecert.c
This is only used from a single place now:
cockpit-session-client-certificate.c.  Delete the file and merge its
code in there.
