+++ src/ws/Makefile-ws.am
@@ -61,8 +61,6 @@ cockpit_session_SOURCES = \
 	src/common/cockpitcloserange.c \
 	src/ws/cockpit-session-client-certificate.h \
 	src/ws/cockpit-session-client-certificate.c \
-	src/ws/cockpitwsinstancecert.h \
-	src/ws/cockpitwsinstancecert.c \
 	src/ws/session-utils.c \
 	src/ws/session-utils.h \
 	src/ws/session.c \
